Commands.txt
Hi Somnath,

I was not able to use the terminal and the SCP command to copy the folder from my local machine to the cloud. I then realised that I have to maybe transfer the zip file and then unzip it. I had to search and find the commands which took some time and digging, so I request you to update it.